refactor: Navbar 구조 개선 및 관심사 분리
