+++ src/types/navigation.ts
@@ -1,0 +1,8 @@
+export type Role = 'guest' | 'manager' | 'president';
+
+export type NavItemData = {
+  key: string;
+  label: string;
+  to?: string;
+  isLogo?: boolean;
+};
